+++ ios/IVSBroadcastCameraView/IVSHelperModule.h
@@ -1,0 +1,9 @@
+//
+//  IVSHelperModule.h
+//  amazon-ivs-react-native-broadcast
+//
+//  Created by Jianlong Nie on 2025/1/13.
+//
+#import <React/RCTBridgeModule.h>
+@interface IVSHelperModule : NSObject <RCTBridgeModule>
+@end
