+++ src/IVSBroadcastCameraView/IVSBroadcastCameraView.types.ts
@@ -1,6 +1,9 @@
 import type { Component, ComponentType } from 'react';
-import type { NativeSyntheticEvent, ViewStyle, StyleProp } from 'react-native';
-import {NativeModules} from "react-native"
+import type { NativeSyntheticEvent, StyleProp, ViewStyle } from 'react-native';
+import { NativeModules } from 'react-native';
+
+console.log('NativeModules=====');
+console.log(NativeModules);
 
 const { IVSHelperModule } = NativeModules;
 
@@ -238,5 +241,14 @@ export interface IIVSBroadcastCameraView {
    */
   swapCamera(): void;
 }
-export const getAvailableDevices: () => Promise<any> =
-  IVSHelperModule.getAvailableDevices;
+
+export const getAvailableDevices = async () => {
+  try {
+    const devices = await IVSHelperModule.getAvailableDevices();
+    console.log('你大妹妹');
+    return devices;
+  } catch (error) {
+    console.error('Error fetching devices:', error);
+    throw error;
+  }
+};
